Skip local name delimiter in last position
Fixes #97
